Provenance: added 'message show provenance [url or file]'
To retrieve the values added in the "provenance" element in a dataschema, at dataset, table or columns level. The cmd generates an object (json dictionary) of all founded values for provenance. This can be used for automating dataprocessing (creating objects in a DB) where the provenance specifiction is used to translate source column names to the specificied column names.