Provenance: get translated cols and non translated cols
In order fully automate the create of a datastructure based upon a dataschema, all the columns are needed.

This adjustment translates the col names if the provenance attribute is defined. If not is gets the orginale version of the column name.

The result is a full column name list witch can be used for automating the datastructure creation.