[PERMISSIONS] add posibility to set DB permissions for single dataset
In order to use the permissions as an operator in Airflow, the permissions logic should take a single dataset schema and set the permissions for the underling objects.

Because data processing in Airflow can make objects to drop, the set permissions are also deleted. DB users are effected by this. One of the solutions is the grant the privileges again. To use it sparingly, the permissions can be set on one single dataset (and it's underling objects).