[4930] add shortnames
The length of tablenames in postgresql is limited to 63 chars.
Some of our datasets are using long table- and fieldnames.

Because we derive the tablenames of joining tables in many-to-many
relations from the table- and fieldnames, we are hitting this
63 chars limit.

Our initial approach of truncating the names also reached its limit
because this now leads to identical tablenames for different joining
tables.

So, we introduce the possibility to define a `shortname` for the
tables and the fields in the amsterdam schema.

This PR implements this new `shortname` phenomenon.

The basic types for amsterdam schema tables and fields
DatasetTableSchema and DatasetFieldSchema were using `id` and `name`
in a somewhat inconsistent manner.

This has been changed. The `id` is now reserved for the original
names of tables and fields in the amsterdam schema. The `name` is equal
to the `id`, unless it is overridden by a `shortname` in the schema.

There was already some work done to prepare for the `shortname`
introduction. This work has been refactored a little, to make it as
side-effect free as possible.