[DS-331] Enable whole slew linting and typing checks.
Ideally we use the same strictness checks across all our projects.
Recently `dataservice-airflow` had these checks enabled and now we have
a fair amount of them enabled for this project. That this is a good
things is apparent in the huge amount of warnings and errors generated.
So many, that we had to temporarily disable some of them. The idea is
to start somewhat small and enable more in due course (see the select
line in .flake8 and the one commented out.)

MyPy's usage could also use a lot more attention. It currently generates
a massive amount of errors. By enabling it in the pre-commit hook we can
gradually start fixing things. Hopefull we arrive at a situation where
we can run MyPy on the entire project.