Added bigint as datatype
Some sources provide us with an own id value. The value size of these source id's can exceed the maxinum of 32 bytes of the standard int(4) datatype in the database. During data load processing data from source into target table, the flow in Airflow can crash. Therefore the is a support added for bigint. Returing a 64 bytes int(8) datatype in the database.