Fix toCamelCase to titlecase words following numbers.
It previously didn't handle titlecasing words following numbers. Eg it
didn't handle `fu33bar` correctly to become `fu33Bar`. Now it does.

In addition I changed the implementation to one based on a regex. A
proof-of-concept regex version originally was 3x faster than the for
loop. To handle additional cases I simply iterated on the
proof-of-concept. It now is 3 µs slower that the for loop based one that
does not handle the `fu33bar` case.