Refactor shortname implementation
The shortnames for tables/fields were used in the Django implemention
in the dynamic generation of the names of the django models.

This led to some weird and hard to debug problems.

The implementation has been changed now. The `table.id` and
`field.id` of the amsterdam schemas are used exclusively throughout Django.

Shortnames (leading to different `table.name` and `field.name` in the
amsterdam schemas are *only* used for the database names.