Add event import for full events
Until now, GOB was only providing partial event records.
GOB is implementing as Kafka producer for full events
(also modify events will contain the full content of a record).

Add fixture and setting to log SQL statements

During testing it can be convenient to see the SQL statements that
are generated by SQLAlchemy. A logging fixture has been added
for the 'sqlalchemy.engine' logger.

To have immediate feedback in the console during tests, the
log_cli setting of pytest can be enabled. It has been added (as
disabled) to the pytest settings in setup.cfg.