[AB#6837] add DB schema name as parameter for table creation
Currently in Airflow an opertor is used in serval dags to create the database objects (tables) that are defined in the dataschema definition (JSON schema). It defaults to the DB schema public. In some dags the tables are first created in a different schema then public before moved to public. Therefor it could be nessacary to create the objects in a different DB schema then public. The DB schema name is added to specify the DB target schema. When not specified it defaults to public.