Fix use of db_table_name
The db_table_name should be passed on as-is to table_factory,
and not modified before that. Because it signals that the
name is passed in or calculated using `table.db_name()`.
In the latter case, shortening of the name should take into
account that a postfix is being involved in the calculation
of the max length of the table name.