AB#4367 Fix MyPy and flake8 errors
This commit was the reason for reinstating MyPy as a pre-commit hook.
The code didn't match the type annotations and could have resulted in a
`AttributeError` if we didn't add the extra guard. This shows the value
that MyPy can bring to a code base. While we were at it we solved the
flake8 errors in this file.