Remove use_dimension_fields flag from datasetschema
The `use_dimension_field` is adding too much magic
and was mainly added for import of partial events
which is now superseded by the import of full events.

Is now obsolete and YAGNI, because it was added
to support partials events which has been ditched in favor of full events.