Fix hardcoded temporal names in index generation
Index creation was failing, because it was using hard-coded names
"identificiatie" and "volgnummer" for the construction
of the temporal fieldnames.
However, these names are dynamic and can change on a per-table basis.

The logic for creating fields in the through tables is already available
in `types.build_through_table`.
So now, the implementation changed to just selecting the appropriate
fields from the dynamically generated through schema.

The algorithm used is:

Get the fields in the dynamic throught tables that are relations
(pointing to the left- and right-side of the relation).

The names of those relation fields are also used to generate some extra
fields in the dynamic through schema (in `types.build_through_table`.
So, a simple `startswith` match can be used to find these extra fields.