Autoload dataset schemas that are not found in the datasets_cache
For relation fields, it is sometimes (esp. for temporal datasets) needed
to load the dataset schemas that are the target for the relation.

For DSO all schemas need to be loaded anyway, so in that case this is
not a problem.

For cli commands, processing one dataset schema, an option (-a) has been added
to load additional dataset when needed.

However, this is a bit cumbersome. And, it does not work for Django
management commands.

The DatasetCollection (holding a cache of datasets) now tries to load
a schema from the SCHEMA_URL location on a cache-miss for a dataset.

There is one caveat for DSO Useage.
Because these schemas are loaded from SCHEMA_URL
it is possible that this schema differs from the schema that
is imported into the PostgreSQL table by the DSO `import_schemas`
command.