Fix the Django classname for a "through" model
The through classname that is automatically generated
should use `field.id` and not `field.name`,
because, when a `shortname` is defined for a field
in the amsterdam schema, this should not affect
the name of the generated through class.

Because, for table fields in amsterdam schema:

`field.id` is always the id in the schema
`field.name` is also the id in the schema,
unless there is a `shortname` defined, in that case
`field.name` is the `shortname`.