Refactor authentication to use native schema objects
Previously, the authentication was implemented in schematools.contrib.django,
which is a high level addition on top of the schema data. This refactor moves
that logic to pure-python classes that read the data directly from the
Amsterdam Schema files + profiles.

This also takes both rulesets into account in a single engine, which
was previously an addon on top of the existing auth.

Some of the changes:

* RequestProfile -> UserScopes
* Django models are no longer read to get "auth" information.
* Django backend uses Existing models to find schema metadata.
* The "auth" and "permissions" properties return parsed data,
  callers no longer have to parse that themselves.
* Graned Permission objects are orderable, and define the level + sub value.
* The Django fields reference their original schema definition, so this
  no longer has to be guessed via to_snake_case() and other comparisons.