AB#4378 Add id column for through tables
The temporal fields for "geldigheid" need to be taken from
the through table. However, to be able to query for these fields
in DSO API, this through table needs to have a non-compound primary key.

This PR adds such a PK field and adjusts the imports to give this
PK field the value of the concatenation of both sides of the relation.