Fix rule to determine loose relationship
When a relation is defined in am. schema,
it can be defined as an object, with > 1 fields.
Those fields (or a subset thereof) are defining
a compound FK to the target of the relationship.

If there is a partial match (for GOB data only a match
on "indentificatie" and not on "volgnummer") the relation
is a so-called "loose" relation.

Because a M2M relation can have additional fields
"on-the-relation" (eg. geldigheid), the fields
that are in the compound relation only need to be
a subset of all the fields in the relation definition.

That check was wrong and has now been fixed.