Add a custom request.Session mocker.
It turned out that `requests.Session` was introduced with
the effect that the mocking of http requests was not working anymore.

The original approach (without requests.Session) was based on the
`requests-mock` package. However it was cumbersome (if not impossible)
to make this work correctly to mock `requests.Session`.

So, a custom duck-type based mock was added to be able to correctly
mock `requests.Session`.