Move factories to a more common location
The `schematools.importer.base` modules had as `table_factory` function
that produces SqlAlchemy models based on amsterdam schema definition.

During the work on events, a more generic `tables_factory` was written
that is having the same purpose, but is more generic and more
functionality has been pushed to the amsterdam schema models in `schematools.types`.

This commit removes the factory function in `schematools.importers.base`
in favor of the new `tables_factory`.