AB 19187 import relations separately using GraphQL API
The GOB datasets (gebieden, bag, brk, meetbouten, hr, woz)
are imported using the GraphQL API provided by GOB.

Relations are imported as embedded objects inside of the main objects.
However, only the active relations are imported, while all the
relations need to be imported, to provide the correct historical data.

Furthermore, the 1-N relations are imported as foreign keys in the
main (parent) objects, so there is no room to store the extra temporal
date (`geldigOp`). This means, the 1-N relations also need to be provided with
so-called through tables, to store this extra information.

The ndjson importer has been changed in such a way that it is now
able to imported data directly in these through tables.

Fieldname mappings are added for the incoming ndjson rows,
Because the GOB names do not follow the amsterdam schema.

Extra fix:

Fix argument name + add optional argument for `generate_db_objects`
The `table_name` is changed to `table_id`. The `table.id` is the
canonical id that is in the amsterdam schema. The `table_name` could
be different when a shortname is defined in the schema.
We really mean the `table.id` in this case.