Swap first two args of _get_dataset_schema.
That way we can make `schema_url` (originally the first param, now the
second param) optional with a default value. That matches the use case
when `schema_location` (orginally the second param, now the first param)
refers to a filename and `schema_url` is not used at all.

Will probably change the `schema_location` parameter name. It is not a
location. It is either an ID of filename. Its name should reflect that.