Add assert to _get_dataset_schema.
With `schema_url` being optional and a rather simple heuristic for
determining whether `schema_location` refers to a file, we run the risk
of incorrectly ending up in the `else` clause. Until we have determined
a better heuristic, add an assert with a hopeful meaningful message for
whoever runs into this.