Deprecate schema_fetch_url_file.
It does not work with datasets that have their tables split out into
separate files. However, it is still used in the permissions_apply CLI
command that I don't know how to fix (what are profiles?). Until I do,
this is the next best thing to discourage it continued use.