Fix the `related_name` for through tables
For the associated table and field the `id` attribute should be used.
The `name` attribute is representing a `shortname`, if available.
In other cases, it is equal to the `id`.
This `shortname` should only be used for database related values.
It is needed, because of the hard limitation on table- and fieldnames
in postgresql.