Add special case to rowclass for provenance with `id` field
When the `id` field in incoming data is used in a provenance,
special care needs to be taken because in some situations
an `id` field is autogenerated to provide the singular PK value
that Django needs.

So, the original value needs to be retained (it would otherwise
be overwritten with the generated `id` field).

And, when asking for the provenanced field, the provenancing
mechanism should *not* to be applied.

So, in this example, the schema fragment is:

	"neuronId": {
		"type": "string",
		"provenance": "id"
	}

The incoming json contains an `id` field that need to become
available as `neuronId`.

Without precautions, the autogenerated `id`, e.g. having a value
of "012.5" would be put in the place of the `neuronId` field
and the original value would be lost.