AB#23066 Add configurable loader for datasets
When a dataset has relations, the datasets that
are targeted in these relations sometimes also need to be loaded,
especially for temporal relations.

There was only one method to load the additional datasets,
using an environment variable `SCHEMA_URL` that was
pointing to a web location with all available amsterdam schema definitions.

However, this is not flexible enough. Especially, when validating a
dataset, more control is needed in fetching related datasets.

This commit adds a configurable loader. By default, the
original method (using `SCHEMA_URL` for http) is used.

A second loader has been added, that is getting related
datasets from the filesystem. This loader is going to
be used for github actions validation of amsterdam schemas.

The `SCHEMA_URL` enviroment variable can be a url
(starting with http(s)) or a filesystem path.
At module load time, the default loader is configured.

When using a cli function, the loader can be reconfigured
by using the `--schema-url` cli option.

Cli commands working with a dataset did have a first
argument that could be a dataset id or a path.
This is a bit confusing, so this has been changed,
the first argument now is always a dataset id.

The value of schema_url (either from an environment
variable or overridden using `--schema-url`) is now
determining where the datasets are located.