explicit database schema default when None
The method tables_factory retrieves database schema if defined as a parameter when creating a database table. If schema is None, it now defaults to Public. Leave it to None will have a risk that the DB schema that is currently in use will be used to create the table in leading to unwanted/unexepected results.