Replace DatasetFieldSchema.sub_fields by subfields (#211)
This is the proper spelling.

More importantly, I wanted to change the semantics of the method and
property, and renaming it seemed the easiest option to break any code
that relied on the old behavior and force reconsideration.

The new method and property return an empty iterator/list for an array
or object instead of raising a ValueError. A use site in DSO-API caught
the ValueError, but as a result also caught unrelated errors, causing
a bug in production and tests successes where failures were appropriate.
It looks like the only other call site performs a custom type check and
should not be affected.

Also changed the return type to match the new code.