AB#23973 Generate JSON properly for DatasetSchema.
Previously, `DatasetSchema`'s internal representation was that of
deserialized Amsterdam Schema JSON (into a Python dict). This internal
representation could be modified by the various `DatasetSchema` methods,
but always retained the primitive types of JSON and as such remained
(deserialized) JSON.

With the introduction of `TableVersions` in commit
1c1191425e11e8ec5b8471b886f31322225cec6f, that introduces a new type of
node somewhere in the tree of dictionaries that DatasetSchema internally
is, that property was lost. One could argue that was a poor choice, but
as the Amsterdam Meta Schema does not specify how multiple active
versions of a given table can be represented inline with the dataset,
this gave us some leeway. Hence instead of messing around with a
`TypedDict` we chose a `dataclass` that has some nice additional
properties.

This does mean though, that we can no longer use `self.data` as a
deserialized JSON representation. Instead we have to do the JSON
dumps/loads dance with a specialized encoder for `TableVersions`.

This will eventually change if we get an updated Amsterdam Meta Schema
definition for multiple active table versions *inline* with the dataset.
But that is not the scope of all these commits. Right now we just try to
get things working within the current active set of
decisions/assumptions/practises around active table versions.