AB#25070 Fix throughtable for loose 1-N relation
Through tables are created For 1-N temporal relations.
Usually, 1-N relations only need an FK in the source table.
However, for temporal relations, addional information about the
temporality needs to be stored in a through tables.

But, for a loose relation, this in not neccessary
and even unwanted. This commit checks for "looseness"
and acts accordingly by not generating through table definitions.

The check for "looseness" that was in the Django model factory
is now needed during schema generation, so it has been moved
to the `types` modules.