AB#25069 Generates correct ids in through table relation
When a through table is generated, two FK relations are inserted in this table,
one to the source and one to the target table.
When dataset.id or table.id in the source or target table are camelCased,
the generated identifiers should stay intact, and should not be snakecased.

This fix also needs some changes in the ndjson importer.
In de ndjson import, the assumption was made that identifiers
were already snakecased. However, that is not correct.
In "schema-land" the camelcasing of identifiers should stay intact.
Only on the interface to the database (e.g. in the ndjson import)
should the neccesary snakecasing be applied to identifiers.