AB#23973 Introduce TableVersions to retain active table versions.
Until the Amsterdam Meta Schema is amended to retain information on
active table versions, we have to track that information by inserting a
data structure in the deserialized JSON that can capture that data:
`TableVersions`

However, even with this commit things are still not well. Although we
now retain the information upon reading the tables from either a path or
URL, we have to ditch it again once we serialize the DatasetSchema back
to JSON in order, for example, to store it in DSO-API's database. Hence
the introduction of `TableVersionsEncoder`.