Upgrade jsonschema to 4.2.1
It has support for newer than `Draft-07` JSON-Schema versions. As it is we
specify `Draft-07` in our schemas even though we are using `Draft
2019-09` features such as `$vocabulary`. Hence we should update the
specification we are referring to sooner rather than later.

Also updated the invocation of the `validate` function. As it turns out
(and now I'm quoting from the `jsonschema` docs:

    JSON Schema defines the `format` property which can be used to check
    if primitive types (`strings`, `numbers`, `booleans`) conform to
    well-defined formats. By default, no validation is enforced, but
    optionally, validation can be enabled by hooking in a
    format-checking object into an `IValidator`.

The updated invocation of `validate` now hooks in a format-checking
object.