AB#23896 A dict is not sufficient to create a DatasetTableSchema.
The `from_dict` method was a weird one to begin with.
`DatasetTableSchema` is-a `UserDict` Hence regular class instantiation
would have sufficed; after all a `UserDict` accepts a `dict`. All
`from_dict` effectively did, was to incorrectly hide any changes in the
class' initialization. That become apparent when `parent_schema` became
a required parameter.

For now we capture calls to `from_dict` just to raise a meaningful error
message.