AB#23896 Add `is_versioned_dataset` property to BaseImporter.
It is a `@cached_property` as it needs to be determined only once for a
given dataset and there is a `BaseImporter` instance per dataset.

As an alternative to querying the DB and employing some heuristics we
could have opted to extend the Amsterdam Schema with an extra property
that specifies how the schema should be represented in the DB. Although
Amsterdam Schema already is a leaky abstraction by virtue of having a
`shortname` and a `provenance` property, that both can be used to
provide details on how (meta)data is going to be stored, such a property
was decided against. After all this is a temporary situation. Some time
in the future all datasets will have a versioned representation in the
DB (with a PostgreSQL schema per dataset) and code is more malleable (in
theory) than a specification such as Amsterdam Schema.