AB#23896 Push versioned table creation into index_factory.
Rather weird that we have to repeat that logic here, but as table names
are determined here we need to extend that logic here as well.

Fixed inner functions from messing with the outer function's namespace.
As result we can do away with the code working around MyPy. MyPy was
simply right.