AB#23896 Fix heuristic to determine whether table has a postfix.
This heuristic is broken. The two facts that:

- we have specified a custom name for a parent table
- we are now processing a derived table

alone are not sufficient to determine we should employ a postfix for our
derived tables as well. A better, but still broken, heuristic is to
additionally check for the presence of the predefined TMP table postfix
in the parent table name.

This is still broken as this only works for tables that use the
predefined postfix for TMP tables. But at least it no longer breaks
versioned table names.