AB#23896 Create VIEWs for versioned tables.
Versioned tables/datasets do live in their own DB schema and the table
names include version numbers. The DSO-API and users of the reference
database do not expect that. So by creating updatable views in the
"public" schema, things should still look the same to them from a usage
perspective.