AB#23896 Extend `schema create all` to work on all datasets.
If no dataset is specified the `schema create all` command will work on
all datasets. Unfortunately this only works for network `schema_url`s
and not filesystem ones.

While testing this command I noticed the `corona` dataset to be in
error; apparently it requires a `crs` property it currently does not
have. Hence I added the ability to exclude datasets from being
processed by means of the `-x` or `--exclude` option. A complete
invocation looks as follows:

    schema create all \
        --db-url postgresql://dataservices:insecure@localhost:5432/dataservices \
        --exclude corona

The `-x` option can be repeated.

This option can also be usefull for proxy datasets.