Rewrite schema validators as decorated functions
All decorators were stateless classes with a single validate method that
were attached to their base class. This made writing new validators a
pain because of all the boilerplate and the flake8 requirement of having
a docstring both on the class and on the validate method.

The new code uses ordinary standalone functions with a decorator to
register them as validators. They're now also all marked as private with
the familiar underscore, and they have a human-readable name that is
independent of their identifier.