[FIX] camelCase in dataset id
The `dataset id` is used for defining the database schema where the tables of that dataset will reaside. In some cases the dataset id is camelCased. Leading to also the prefix of tables - that contain the name of its schema - to have camelCased prefixes. Which leads to wrong table names and issues for the datateams. To cope with this, the dataset id is snaked cased.