Update schema command description in README.md
The list of schema subcommands was out of date. Refer the user to --help
for the current list, so we don't have to maintain it in the README.

The PostgreSQL is now explicitly stated.

Describe how schemas are (usually) loaded. Having a single file with a
dataset schema is not enough for some (most?) commands.