Fix problem with imports of nested tables
The import of temporal relation was split out in separate DAGs recently.
However, in some situation the so-called nested tables are being used.
Esp. in `bag_verblijfsobjecten` for `gebruiksdoel`.

The factory function that is creating the SQLAlchemy models
was not generating models for those nested tables anymore.

Due to that, imports of embedded json objects in the ndjson
was silently skipped.

This has been fixed in this PR.