Recognize the implicit scope "OPENBAAR"
This special scope is defined by the Amsterdam Schema spec, so it makes
sense to handle it entirely in schema-tools rather than DSO-API.

The string constant is now repeated in several modules, because defining
it centrally in permissions.auth causes a circular import (even when in
a submodule permissions.auth._public). Leaving that problem for later,
when we implement closed-by-default logic.