AB#28435 Fix shortnamed identifiers for ndjson import
When a table identifier in an amsterdam schema definition
does have a shortname, the ndjson import is failing.

This commit fixes this behaviour by modifying incoming
records for those shortnamed identifiers.

The idea is that in the incoming ndjson the identifiers
that are used are always according to the amsterdam schema.
The shortname is a implementation detail, that should only
be used at the moment that records are prepared for
insertion into the sql database.