AB#28436 Add permissions for relations
Permissions for relations were not applied at all.

This PR applies the correct permission to the FK field in the base
table. The FK fields do have a `_id` postfix in the postgres db.

On the `DatasetFieldSchema` a `db_name()` has been added, analogous
to the same method on `DatasetTableSchema`.
This `db_name()` method is used to find the correct columns that
needs to have a permission applied.

Permissions for "composite" foreign keys are being added.

Those are not real composite keys. The key column itself
is a single FK, however, for relations to temporary
tables, the key has two dot-separated parts,
the `identificatie` and the `volgnummer`.

Those fields are also stored separately in database columns.
So, those columns also need to be protected with permissions.

Furthermore permissions for through tables and nested tables
are added.

For both through and nested relations (type `array`)
a separate table is being used in postgresql.

For those tables, no grants were being set. However,
those tables need to be protected according to the scope (`auth` field)
that is on the relation in the amsterdam schema.

The current implementation was already very bloated and
totally not DRY and could not accommodate this change.

So the implementation has been refactored from the ground up.