AB#42523 Add model mocker factory
First version of a model mocker.
The mocking is done based on the amsterdam schema, and the
dynamic Mock model is created in a similar fashion as the regular Django
models.

Fakers
======

For the models fields, fakers are created based on the type
of the field in the amsterdam schema. So, for a `string`
fields, a random string is generated as mock data.

If a `format` is availble in the definition (e.g. for a `date-time`)
this information is used to look up the appropriate faker.

Furthermore, a `faker` attribute can be added to the field
definition. This field should be the unique name of a faker.
Be it one of the provided fakers (eg `ssn` faker for the dutch `bsn`)
or a custom built faker.

Maybe we want to change the name of this `faker` attribute to `format`,
however, that cannot be done for now, it would crash the DSO API.

Custom fakers
=============

Two custom fakers are added. One is `int_window` that produces
an integer in a certain range. This is a simple faker, added as an
example.

The second faker is the `geo` faker. This faker can produce data for one
of the standard geo types (e.g. Point, MultiPoint, Polygon etc.).

This geo faker is still very limited and needs a lot of enhancements.

- The CRS is now uses incorrectly
- We should come up with better shapes

Lots of amsterdam schemas are now using the `Geometry` type.
However, this is an abstract type, so not data can be produced
for this type. So, schemas need to be changed to use not-abstract types.

If the `Geometry` type is used, a `null` record is created.

Relations
=========

Before the model is built, the schema is tweaked in such a way
that relations are replaced by simple scalar fields with an `_id`
postfix. Otherwise, during mocking there is a huge graph of
interconnected models that needs to generated in exactly the
correct order (sometimes even circular patterns in the graph).

The idea is to create a second step in the process that can change
the data and add relation. A bit in a similar fashion as BenK
doing a separete "relate" step. However, this idea is still
rough around the edges at the moment.

Generating data
===============

To generate the data, an django management command `create_mock_data` has been added.

This mgm. command works in a similar fashion as the `import_schemas`
mgm. command. Without parameters, it will try to generate mock data for
all datasets. Alternatively,  the path to one or more schema files can be given, to
create mock data for just these schemas.

NB. The tables should already have been created in the database.