[AB#50108] Convert LooseRelationField into an actual foreign key object.
This replaces the separate `CompositeForeignKeySubField` field,
as both are now exactly the same.

This adds a backwards incompatibility for reading the field,
as it now returns the relation object instead of the ID value.
Reading the field will likely result in an MultipleObjectsReturned
exception, but joins will work fine, thanks to the `to_field` value.