Introduce new loader, and remove singleton _DatasetCollection class
This loader replaces the logic in `schematools.utils`. Instead of having
separate functions, there is a single interface that can be implemented
for every type of loading (URL, filesystem or database or from cache).

This interface also allows retrieval of other table versions
(`loader.get_table()`), which is the main motivation for this change.

Secondary benefit is that the `_DatasetCollection` could be removed.
Instead of having a global cache, the datasets are only cached within
the same loader instance (or `QuerySet`) that retrieves the data.

However, this will break code that assume datasets are globally cached.
This affected both the Django management commands, and unit tests.
The tests use a shared `schema_loader` fixture instead, so all schemas
from the same test are cached.