Cleanup and simplify model field factory logic
This replaces the complex FieldMaker/RelationMaker inheritance logic,
with all its back-and-forth jumps, with plain Python functions.

This also accidentally fixes a bug with geometry fields.
Because `JSON_TYPE_TO_DJANGO` has a a static kwargs argument for those,
it would be directly altered by the `FieldMaker` - and therefore the
next could field receive the `verbose_name` of the previous last
geometry field that had one.