AB#65251 Fix mutation issue in mocker declaration
For the FactoryBoy Declarations (the mockers) a callable wrapper class is
being created to hold the actual FactoryBoy Declaration and also the
arg and kwargs that are needed during mocking.

These wrappers are declared as class instances at the module level. However,
these instances were being mutated, so state was leaking from one
call to the next call.

This PR fixes this issue.